chore: modularize protocol and standardize parity tests (Issue #18)
- Decomposed three large protocol files into domain-focused submodules:
  - game_message.rs shifted to messages/game_message/
  - object/messages.rs shifted to messages/object/messages/
  - movement/messages.rs shifted to messages/movement/messages/
- Standardized parity tests in character, inventory, and network modules using the shared assert_pack_unpack_parity helper.
- Established a canonical fixture provenance workflow in FIXTURES.md and PROVENANCE.md.
- Updated documentation and worksheets to reflect the new modular structure.
- Fixed clippy lints and applied cargo fmt across the workspace.
