+++ crates/holtburger-protocol/src/messages/game_message/tests.rs
@@ -1,0 +1,265 @@
+use super::GameMessage;
+use crate::test_fixtures;
+use holtburger_common::traits::{ProtocolPack, ProtocolUnpack};
+
+fn assert_dispatch_match(data: &[u8], check: impl Fn(&GameMessage) -> bool) {
+    let mut offset = 0;
+    let msg = GameMessage::unpack(data, &mut offset).expect("Failed to unpack GameMessage");
+    if !check(&msg) {
+        panic!("Dispatch failed. Got: {:?}", msg);
+    }
+
+    // Also verify pack parity for these
+    let mut packed = Vec::new();
+    msg.pack(&mut packed);
+    assert_eq!(packed, data, "Pack parity failed");
+}
+
+fn assert_dispatch_match_no_parity(data: &[u8], check: impl Fn(&GameMessage) -> bool) {
+    let mut offset = 0;
+    let msg = GameMessage::unpack(data, &mut offset).expect("Failed to unpack GameMessage");
+    if !check(&msg) {
+        panic!("Dispatch failed. Got: {:?}", msg);
+    }
+}
+
+#[test]
+fn test_dispatch_character_list() {
+    assert_dispatch_match(test_fixtures::CHARACTER_LIST, |msg| {
+        matches!(msg, GameMessage::CharacterList(_))
+    });
+}
+
+#[test]
+fn test_dispatch_character_enter_world() {
+    assert_dispatch_match(test_fixtures::CHARACTER_ENTER_WORLD, |msg| {
+        matches!(msg, GameMessage::CharacterEnterWorld(_))
+    });
+}
+
+#[test]
+fn test_dispatch_character_enter_world_request() {
+    assert_dispatch_match(test_fixtures::CHARACTER_ENTER_WORLD_REQUEST, |msg| {
+        matches!(msg, GameMessage::CharacterEnterWorldRequest(_))
+    });
+}
+
+#[test]
+fn test_dispatch_play_sound() {
+    assert_dispatch_match(test_fixtures::SOUND, |msg| {
+        matches!(msg, GameMessage::PlaySound(_))
+    });
+}
+
+#[test]
+fn test_dispatch_play_effect() {
+    assert_dispatch_match(test_fixtures::PLAY_EFFECT, |msg| {
+        matches!(msg, GameMessage::PlayEffect(_))
+    });
+}
+
+#[test]
+fn test_dispatch_hear_speech() {
+    assert_dispatch_match(test_fixtures::HEAR_SPEECH, |msg| {
+        matches!(msg, GameMessage::HearSpeech(_))
+    });
+}
+
+#[test]
+fn test_dispatch_character_error() {
+    // WeenieError is wrapped in GameEvent, not top-level CharacterError
+    assert_dispatch_match(test_fixtures::WEENIE_ERROR, |msg| {
+        matches!(msg, GameMessage::GameEvent(_))
+    });
+}
+
+#[test]
+fn test_dispatch_ddd_interrogation_response() {
+    assert_dispatch_match(test_fixtures::DDD_INTERROGATION_RESPONSE, |msg| {
+        matches!(msg, GameMessage::DddInterrogationResponse(_))
+    });
+}
+
+#[test]
+fn test_dispatch_game_action_move_to_state() {
+    assert_dispatch_match(test_fixtures::MOVE_TO_STATE, |msg| {
+        matches!(msg, GameMessage::GameAction(_))
+    });
+}
+
+#[test]
+fn test_dispatch_player_teleport() {
+    assert_dispatch_match(test_fixtures::PLAYER_TELEPORT, |msg| {
+        matches!(msg, GameMessage::PlayerTeleport(_))
+    });
+}
+
+#[test]
+fn test_dispatch_game_event_player_description() {
+    // KNOWN PARITY GAP: PlayerDescription is currently a special case from the Gold Standard
+    // because it contains complex nested structures (like EnchantmentRegistry) that we
+    // haven't fully implemented bit-perfect packing for yet. Additionally, GameplayOptions
+    // are extracted via a heuristic which can lead to repacking drift.
+    assert_dispatch_match_no_parity(test_fixtures::PLAYER_DESCRIPTION, |msg| {
+        matches!(msg, GameMessage::GameEvent(_))
+    });
+}
+
+#[test]
+fn test_dispatch_action_talk() {
+    assert_dispatch_match(test_fixtures::ACTION_TALK, |msg| {
+        matches!(msg, GameMessage::GameAction(_))
+    });
+}
+
+#[test]
+fn test_dispatch_action_tell() {
+    assert_dispatch_match(test_fixtures::ACTION_TELL, |msg| {
+        matches!(msg, GameMessage::GameAction(_))
+    });
+}
+
+#[test]
+fn test_dispatch_action_use() {
+    assert_dispatch_match(test_fixtures::ACTION_USE, |msg| {
+        matches!(msg, GameMessage::GameAction(_))
+    });
+}
+
+#[test]
+fn test_dispatch_action_drop_item() {
+    assert_dispatch_match(test_fixtures::ACTION_DROP_ITEM, |msg| {
+        matches!(msg, GameMessage::GameAction(_))
+    });
+}
+
+#[test]
+fn test_dispatch_action_put_item() {
+    assert_dispatch_match(test_fixtures::ACTION_PUT_ITEM, |msg| {
+        matches!(msg, GameMessage::GameAction(_))
+    });
+}
+
+#[test]
+fn test_dispatch_action_identify() {
+    assert_dispatch_match(test_fixtures::ACTION_IDENTIFY, |msg| {
+        matches!(msg, GameMessage::GameAction(_))
+    });
+}
+
+#[test]
+fn test_dispatch_action_login_complete() {
+    assert_dispatch_match(test_fixtures::ACTION_LOGIN_COMPLETE, |msg| {
+        matches!(msg, GameMessage::GameAction(_))
+    });
+}
+
+#[test]
+fn test_dispatch_action_ping_request() {
+    assert_dispatch_match(test_fixtures::ACTION_PING_REQUEST, |msg| {
+        matches!(msg, GameMessage::GameAction(_))
+    });
+}
+
+#[test]
+fn test_dispatch_private_update_position() {
+    assert_dispatch_match(test_fixtures::PRIVATE_UPDATE_POSITION, |msg| {
+        matches!(msg, GameMessage::PrivateUpdatePosition(_))
+    });
+}
+
+#[test]
+fn test_dispatch_public_update_position() {
+    assert_dispatch_match(test_fixtures::PUBLIC_UPDATE_POSITION, |msg| {
+        matches!(msg, GameMessage::PublicUpdatePosition(_))
+    });
+}
+
+#[test]
+fn test_dispatch_hear_ranged_speech() {
+    assert_dispatch_match(test_fixtures::HEAR_RANGED_SPEECH, |msg| {
+        matches!(msg, GameMessage::HearRangedSpeech(_))
+    });
+}
+
+#[test]
+fn test_dispatch_emote_text() {
+    assert_dispatch_match(test_fixtures::EMOTE_TEXT, |msg| {
+        matches!(msg, GameMessage::EmoteText(_))
+    });
+}
+
+#[test]
+fn test_dispatch_soul_emote() {
+    assert_dispatch_match(test_fixtures::SOUL_EMOTE, |msg| {
+        matches!(msg, GameMessage::SoulEmote(_))
+    });
+}
+
+#[test]
+fn test_dispatch_force_obj_desc_send() {
+    assert_dispatch_match(test_fixtures::FORCE_OBJ_DESC_SEND, |msg| {
+        matches!(msg, GameMessage::ForceObjectDescSend(_))
+    });
+}
+
+#[test]
+fn test_dispatch_obj_desc_event() {
+    assert_dispatch_match(test_fixtures::OBJ_DESC_EVENT, |msg| {
+        matches!(msg, GameMessage::ObjDescEvent(_))
+    });
+}
+
+#[test]
+fn test_dispatch_update_skill_level_private() {
+    assert_dispatch_match(test_fixtures::UPDATE_SKILL_LEVEL_PRIVATE, |msg| {
+        matches!(msg, GameMessage::PrivateUpdateSkillLevel(_))
+    });
+}
+
+#[test]
+fn test_dispatch_update_skill_level_public() {
+    assert_dispatch_match(test_fixtures::UPDATE_SKILL_LEVEL_PUBLIC, |msg| {
+        matches!(msg, GameMessage::PublicUpdateSkillLevel(_))
+    });
+}
+
+#[test]
+fn test_dispatch_view_contents() {
+    assert_dispatch_match(test_fixtures::VIEW_CONTENTS, |msg| {
+        matches!(msg, GameMessage::GameEvent(_))
+    });
+}
+
+#[test]
+fn test_dispatch_weenie_error_with_string() {
+    assert_dispatch_match(test_fixtures::WEENIE_ERROR_WITH_STRING, |msg| {
+        matches!(msg, GameMessage::GameEvent(_))
+    });
+}
+
+#[test]
+fn test_dispatch_update_property_int_event() {
+    assert_dispatch_match(test_fixtures::UPDATE_PROPERTY_INT, |msg| {
+        matches!(msg, GameMessage::GameEvent(_))
+    });
+}
+
+#[test]
+fn test_private_update_combat_mode_parity() {
+    let fixture = hex::decode("CD0200000C2800000002000000").unwrap();
+    let mut offset = 0;
+    let msg = GameMessage::unpack(&fixture, &mut offset).expect("failed to unpack GameMessage");
+
+    if let GameMessage::PrivateUpdatePropertyInt(ref data) = msg {
+        assert_eq!(data.sequence, 0x0C);
+        assert_eq!(data.property, 40); // CombatMode
+        assert_eq!(data.value, 2); // Melee
+    } else {
+        panic!("expected PrivateUpdatePropertyInt, got {:?}", msg);
+    }
+
+    let mut packed = Vec::new();
+    msg.pack(&mut packed);
+    assert_eq!(packed, fixture);
+}
