+++ crates/holtburger-protocol/src/messages/movement/messages/motion.rs
@@ -1,0 +1,409 @@
+use crate::messages::movement::types::*;
+use crate::messages::utils::{align_offset, pad_to_4};
+use byteorder::{ByteOrder, LittleEndian};
+use holtburger_common::Guid;
+pub use holtburger_common::position::WorldPosition;
+use holtburger_common::traits::{ProtocolPack, ProtocolUnpack};
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct MovementEventData {
+    pub guid: Guid,
+    pub object_instance_sequence: u16,
+    pub movement_sequence: u16,
+    pub server_control_sequence: u16,
+    pub is_autonomous: bool,
+    pub movement_type: MovementType,
+    pub motion_flags: u8,
+    pub current_style: u16,
+    pub data: MovementTypeData,
+}
+
+impl ProtocolUnpack for MovementEventData {
+    fn unpack(data: &[u8], offset: &mut usize) -> Option<Self> {
+        let guid = Guid::unpack(data, offset)?;
+
+        if *offset + 2 > data.len() {
+            return None;
+        }
+        let object_instance_sequence = LittleEndian::read_u16(&data[*offset..*offset + 2]);
+        *offset += 2;
+
+        if *offset + 2 > data.len() {
+            return None;
+        }
+        let movement_sequence = LittleEndian::read_u16(&data[*offset..*offset + 2]);
+        *offset += 2;
+
+        if *offset + 2 > data.len() {
+            return None;
+        }
+        let server_control_sequence = LittleEndian::read_u16(&data[*offset..*offset + 2]);
+        *offset += 2;
+
+        if *offset + 1 > data.len() {
+            return None;
+        }
+        let is_autonomous = data[*offset] != 0;
+        *offset += 1;
+
+        // Alignment (ACE uses Writer.Align() which aligns to 4 bytes)
+        align_offset(offset, 4);
+
+        if *offset + 1 > data.len() {
+            return None;
+        }
+        let movement_type_raw = data[*offset];
+        let movement_type =
+            MovementType::from_repr(movement_type_raw).unwrap_or(MovementType::Invalid);
+        *offset += 1;
+
+        if *offset + 1 > data.len() {
+            return None;
+        }
+        let motion_flags = data[*offset];
+        *offset += 1;
+
+        if *offset + 2 > data.len() {
+            return None;
+        }
+        let current_style = LittleEndian::read_u16(&data[*offset..*offset + 2]);
+        *offset += 2;
+
+        let data_payload = match movement_type {
+            MovementType::MoveToObject => {
+                MovementTypeData::MoveToObject(MoveToObject::unpack(data, offset)?)
+            }
+            MovementType::MoveToPosition => {
+                MovementTypeData::MoveToPosition(MoveToPosition::unpack(data, offset)?)
+            }
+            MovementType::TurnToObject => {
+                MovementTypeData::TurnToObject(TurnToObject::unpack(data, offset)?)
+            }
+            MovementType::TurnToHeading => {
+                MovementTypeData::TurnToHeading(TurnToHeading::unpack(data, offset)?)
+            }
+            MovementType::Invalid
+            | MovementType::RawCommand
+            | MovementType::InterpretedCommand
+            | MovementType::StopRawCommand
+            | MovementType::StopInterpretedCommand
+            | MovementType::StopCompletely => {
+                MovementTypeData::Invalid(MovementInvalid::unpack_ext(data, offset, motion_flags)?)
+            }
+        };
+
+        Some(MovementEventData {
+            guid,
+            object_instance_sequence,
+            movement_sequence,
+            server_control_sequence,
+            is_autonomous,
+            movement_type,
+            motion_flags,
+            current_style,
+            data: data_payload,
+        })
+    }
+}
+
+impl ProtocolPack for MovementEventData {
+    fn pack(&self, buf: &mut Vec<u8>) {
+        self.guid.pack(buf);
+        buf.extend_from_slice(&self.object_instance_sequence.to_le_bytes());
+        buf.extend_from_slice(&self.movement_sequence.to_le_bytes());
+        buf.extend_from_slice(&self.server_control_sequence.to_le_bytes());
+        buf.push(self.is_autonomous as u8);
+
+        // Alignment
+        pad_to_4(buf);
+
+        buf.push(self.movement_type as u8);
+        buf.push(self.motion_flags);
+        buf.extend_from_slice(&self.current_style.to_le_bytes());
+
+        match &self.data {
+            MovementTypeData::Invalid(d) => d.pack(buf),
+            MovementTypeData::MoveToObject(d) => d.pack(buf),
+            MovementTypeData::MoveToPosition(d) => d.pack(buf),
+            MovementTypeData::TurnToObject(d) => d.pack(buf),
+            MovementTypeData::TurnToHeading(d) => d.pack(buf),
+        }
+    }
+}
+
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub enum MovementTypeData {
+    Invalid(MovementInvalid),
+    MoveToObject(MoveToObject),
+    MoveToPosition(MoveToPosition),
+    TurnToObject(TurnToObject),
+    TurnToHeading(TurnToHeading),
+}
+
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize, Default)]
+pub struct MovementInvalid {
+    pub state: InterpretedMotionState,
+    pub sticky_object: Option<Guid>,
+}
+
+impl MovementInvalid {
+    pub fn unpack_ext(data: &[u8], offset: &mut usize, flags: u8) -> Option<Self> {
+        let state = InterpretedMotionState::unpack(data, offset)?;
+        let sticky_object = if (flags & 0x01) != 0 {
+            Guid::unpack(data, offset)
+        } else {
+            None
+        };
+        Some(MovementInvalid {
+            state,
+            sticky_object,
+        })
+    }
+}
+
+impl ProtocolUnpack for MovementInvalid {
+    fn unpack(data: &[u8], offset: &mut usize) -> Option<Self> {
+        Self::unpack_ext(data, offset, 0)
+    }
+}
+
+impl ProtocolPack for MovementInvalid {
+    fn pack(&self, buf: &mut Vec<u8>) {
+        self.state.pack(buf);
+        if let Some(guid) = self.sticky_object {
+            guid.pack(buf);
+        }
+    }
+}
+
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize, Default)]
+pub struct MoveToObject {
+    pub target: Guid,
+    pub origin: Origin,
+    pub params: MoveToParameters,
+    pub run_rate: f32,
+}
+
+impl ProtocolUnpack for MoveToObject {
+    fn unpack(data: &[u8], offset: &mut usize) -> Option<Self> {
+        let target = Guid::unpack(data, offset)?;
+        let origin = Origin::unpack(data, offset)?;
+        let params = MoveToParameters::unpack(data, offset)?;
+        if *offset + 4 > data.len() {
+            return None;
+        }
+        let run_rate = LittleEndian::read_f32(&data[*offset..*offset + 4]);
+        *offset += 4;
+        Some(MoveToObject {
+            target,
+            origin,
+            params,
+            run_rate,
+        })
+    }
+}
+
+impl ProtocolPack for MoveToObject {
+    fn pack(&self, buf: &mut Vec<u8>) {
+        self.target.pack(buf);
+        self.origin.pack(buf);
+        self.params.pack(buf);
+        buf.extend_from_slice(&self.run_rate.to_le_bytes());
+    }
+}
+
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize, Default)]
+pub struct MoveToPosition {
+    pub origin: Origin,
+    pub params: MoveToParameters,
+    pub run_rate: f32,
+}
+
+impl ProtocolUnpack for MoveToPosition {
+    fn unpack(data: &[u8], offset: &mut usize) -> Option<Self> {
+        let origin = Origin::unpack(data, offset)?;
+        let params = MoveToParameters::unpack(data, offset)?;
+        if *offset + 4 > data.len() {
+            return None;
+        }
+        let run_rate = LittleEndian::read_f32(&data[*offset..*offset + 4]);
+        *offset += 4;
+        Some(MoveToPosition {
+            origin,
+            params,
+            run_rate,
+        })
+    }
+}
+
+impl ProtocolPack for MoveToPosition {
+    fn pack(&self, buf: &mut Vec<u8>) {
+        self.origin.pack(buf);
+        self.params.pack(buf);
+        buf.extend_from_slice(&self.run_rate.to_le_bytes());
+    }
+}
+
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize, Default)]
+pub struct TurnToObject {
+    pub target: Guid,
+    pub desired_heading: f32,
+    pub params: TurnToParameters,
+}
+
+impl ProtocolUnpack for TurnToObject {
+    fn unpack(data: &[u8], offset: &mut usize) -> Option<Self> {
+        let target = Guid::unpack(data, offset)?;
+        if *offset + 4 > data.len() {
+            return None;
+        }
+        let desired_heading = LittleEndian::read_f32(&data[*offset..*offset + 4]);
+        *offset += 4;
+        let params = TurnToParameters::unpack(data, offset)?;
+        Some(TurnToObject {
+            target,
+            desired_heading,
+            params,
+        })
+    }
+}
+
+impl ProtocolPack for TurnToObject {
+    fn pack(&self, buf: &mut Vec<u8>) {
+        self.target.pack(buf);
+        buf.extend_from_slice(&self.desired_heading.to_le_bytes());
+        self.params.pack(buf);
+    }
+}
+
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize, Default)]
+pub struct TurnToHeading {
+    pub params: TurnToParameters,
+}
+
+impl ProtocolUnpack for TurnToHeading {
+    fn unpack(data: &[u8], offset: &mut usize) -> Option<Self> {
+        let params = TurnToParameters::unpack(data, offset)?;
+        Some(TurnToHeading { params })
+    }
+}
+
+impl ProtocolPack for TurnToHeading {
+    fn pack(&self, buf: &mut Vec<u8>) {
+        self.params.pack(buf);
+    }
+}
+
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize, Default)]
+pub struct Origin {
+    pub cell_id: Guid,
+    pub position: holtburger_common::math::Vector3,
+}
+
+impl ProtocolUnpack for Origin {
+    fn unpack(data: &[u8], offset: &mut usize) -> Option<Self> {
+        if *offset + 16 > data.len() {
+            return None;
+        }
+        let cell_id = LittleEndian::read_u32(&data[*offset..*offset + 4]).into();
+        *offset += 4;
+        let x = LittleEndian::read_f32(&data[*offset..*offset + 4]);
+        let y = LittleEndian::read_f32(&data[*offset + 4..*offset + 8]);
+        let z = LittleEndian::read_f32(&data[*offset + 8..*offset + 12]);
+        *offset += 12;
+        Some(Origin {
+            cell_id,
+            position: holtburger_common::math::Vector3 { x, y, z },
+        })
+    }
+}
+
+impl ProtocolPack for Origin {
+    fn pack(&self, buf: &mut Vec<u8>) {
+        buf.extend_from_slice(&<Guid as Into<u32>>::into(self.cell_id).to_le_bytes());
+        buf.extend_from_slice(&self.position.x.to_le_bytes());
+        buf.extend_from_slice(&self.position.y.to_le_bytes());
+        buf.extend_from_slice(&self.position.z.to_le_bytes());
+    }
+}
+
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize, Default)]
+pub struct MoveToParameters {
+    pub movement_parameters: u32,
+    pub distance_to_object: f32,
+    pub min_distance: f32,
+    pub fail_distance: f32,
+    pub speed: f32,
+    pub walk_run_threshold: f32,
+    pub desired_heading: f32,
+}
+
+impl ProtocolUnpack for MoveToParameters {
+    fn unpack(data: &[u8], offset: &mut usize) -> Option<Self> {
+        if *offset + 28 > data.len() {
+            return None;
+        }
+        let movement_parameters = LittleEndian::read_u32(&data[*offset..*offset + 4]);
+        let distance_to_object = LittleEndian::read_f32(&data[*offset + 4..*offset + 8]);
+        let min_distance = LittleEndian::read_f32(&data[*offset + 8..*offset + 12]);
+        let fail_distance = LittleEndian::read_f32(&data[*offset + 12..*offset + 16]);
+        let speed = LittleEndian::read_f32(&data[*offset + 16..*offset + 20]);
+        let walk_run_threshold = LittleEndian::read_f32(&data[*offset + 20..*offset + 24]);
+        let desired_heading = LittleEndian::read_f32(&data[*offset + 24..*offset + 28]);
+        *offset += 28;
+        Some(MoveToParameters {
+            movement_parameters,
+            distance_to_object,
+            min_distance,
+            fail_distance,
+            speed,
+            walk_run_threshold,
+            desired_heading,
+        })
+    }
+}
+
+impl ProtocolPack for MoveToParameters {
+    fn pack(&self, buf: &mut Vec<u8>) {
+        buf.extend_from_slice(&self.movement_parameters.to_le_bytes());
+        buf.extend_from_slice(&self.distance_to_object.to_le_bytes());
+        buf.extend_from_slice(&self.min_distance.to_le_bytes());
+        buf.extend_from_slice(&self.fail_distance.to_le_bytes());
+        buf.extend_from_slice(&self.speed.to_le_bytes());
+        buf.extend_from_slice(&self.walk_run_threshold.to_le_bytes());
+        buf.extend_from_slice(&self.desired_heading.to_le_bytes());
+    }
+}
+
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize, Default)]
+pub struct TurnToParameters {
+    pub movement_parameters: u32,
+    pub speed: f32,
+    pub desired_heading: f32,
+}
+
+impl ProtocolUnpack for TurnToParameters {
+    fn unpack(data: &[u8], offset: &mut usize) -> Option<Self> {
+        if *offset + 12 > data.len() {
+            return None;
+        }
+        let movement_parameters = LittleEndian::read_u32(&data[*offset..*offset + 4]);
+        let speed = LittleEndian::read_f32(&data[*offset + 4..*offset + 8]);
+        let desired_heading = LittleEndian::read_f32(&data[*offset + 8..*offset + 12]);
+        *offset += 12;
+        Some(TurnToParameters {
+            movement_parameters,
+            speed,
+            desired_heading,
+        })
+    }
+}
+
+impl ProtocolPack for TurnToParameters {
+    fn pack(&self, buf: &mut Vec<u8>) {
+        buf.extend_from_slice(&self.movement_parameters.to_le_bytes());
+        buf.extend_from_slice(&self.speed.to_le_bytes());
+        buf.extend_from_slice(&self.desired_heading.to_le_bytes());
+    }
+}
