+++ crates/holtburger-protocol/src/messages/object/messages/mod.rs
@@ -1,0 +1,31 @@
+use holtburger_common::Guid;
+use holtburger_common::traits::{ProtocolPack, ProtocolUnpack};
+
+pub mod attribute;
+pub mod description;
+pub mod properties;
+pub mod sound;
+mod tests;
+
+pub use self::attribute::*;
+pub use self::description::*;
+pub use self::properties::*;
+pub use self::sound::*;
+
+#[derive(Debug, Clone, PartialEq)]
+pub struct ObjectDeleteData {
+    pub guid: Guid,
+}
+
+impl ProtocolUnpack for ObjectDeleteData {
+    fn unpack(data: &[u8], offset: &mut usize) -> Option<Self> {
+        let guid = Guid::unpack(data, offset)?;
+        Some(ObjectDeleteData { guid })
+    }
+}
+
+impl ProtocolPack for ObjectDeleteData {
+    fn pack(&self, buf: &mut Vec<u8>) {
+        self.guid.pack(buf);
+    }
+}
